PBJVision: add a type cast for precision loss of setting verts

+++ Source/PBJVision.m
@@ -1888,14 +1888,14 @@ typedef void (^PBJVisionBlock)();
 
     static GLfloat vertices[8];
 
-    vertices[0] = -widthScale;
-    vertices[1] = -heightScale;
-    vertices[2] = widthScale;
-    vertices[3] = -heightScale;
-    vertices[4] = -widthScale;
-    vertices[5] = heightScale;
-    vertices[6] = widthScale;
-    vertices[7] = heightScale;
+    vertices[0] = (GLfloat) -widthScale;
+    vertices[1] = (GLfloat) -heightScale;
+    vertices[2] = (GLfloat) widthScale;
+    vertices[3] = (GLfloat) -heightScale;
+    vertices[4] = (GLfloat) -widthScale;
+    vertices[5] = (GLfloat) heightScale;
+    vertices[6] = (GLfloat) widthScale;
+    vertices[7] = (GLfloat) heightScale;
 
     static const GLfloat textureCoordinates[] = {
         0.0f, 1.0f,
